+++ gaze.h
@@ -19,7 +19,7 @@ public:
     int left_validation, right_validation,
         source_file_line = -1, source_file_col = -1;
     long long event_time, system_time;
-    float x, y, left_pupil_diameter, right_pupil_diameter;
+    double x, y, left_pupil_diameter, right_pupil_diameter;
     QString gaze_target = "", gaze_target_type = "",
                 source_token = "", source_token_xpath = "", source_token_syntatic_context = "";
     bool initialized = false;
